// frontend/src/generated/SeriesGroup.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SeriesGroupDimension } from './SeriesGroupDimension'
import type { SeriesPoint } from './SeriesPoint'

export type SeriesGroup = {
  dimension: SeriesGroupDimension
  id: string
  label: string
  series: Array<SeriesPoint>
}